Machine learning (model deployment) for Wisconsin Breast Cancer dataset
The dataset used in this project is the Wisconsin Breast Cancer dataset, which consists of various features extracted from breast mass samples. The goal is to classify the samples as either malignant (M) or benign (B) based on these features.
Trained different classifiers such as Logistic Regression, Decision Trees, Random Forests, and SVM on the preprocessed dataset. Evaluated their performance using appropriate metrics and selected the best-performing model. Deploy the selected model on the test dataset and generate predictions